Refactor: Move converter functions to service layer
Move conversion functions from domain models to service layer to maintain
clean architecture boundaries. Domain models should remain pure data
structures without conversion logic.
