+++ internal/service/project_operations.go
@@ -48,7 +48,7 @@ func (s *ProjectsService) GetProjects(ctx context.Context) ([]*projsvc.ProjectFu
 	projectsFull := []*projsvc.ProjectFull{}
 	for uid, projectBase := range projectsBaseMap {
 		projectSettings := projectsSettingsMap[uid] // May be nil if settings don't exist
-		projectFull := models.ConvertToProjectFull(projectBase, projectSettings)
+		projectFull := ConvertToProjectFull(projectBase, projectSettings)
 		if projectFull != nil {
 			projectsFull = append(projectsFull, projectFull)
 		}
@@ -127,13 +127,13 @@ func (s *ProjectsService) CreateProject(ctx context.Context, payload *projsvc.Cr
 		MeetingCoordinators: payload.MeetingCoordinators,
 	}
 
-	projectDB, err := models.ConvertToDBProjectBase(project)
+	projectDB, err := ConvertToDBProjectBase(project)
 	if err != nil {
 		slog.ErrorContext(ctx, "error converting project to DB project", constants.ErrKey, err)
 		return nil, domain.ErrInternal
 	}
 
-	projectSettingsDB, err := models.ConvertToDBProjectSettings(projectSettings)
+	projectSettingsDB, err := ConvertToDBProjectSettings(projectSettings)
 	if err != nil {
 		slog.ErrorContext(ctx, "error converting project settings to DB project settings", constants.ErrKey, err)
 		return nil, domain.ErrInternal
@@ -173,7 +173,7 @@ func (s *ProjectsService) CreateProject(ctx context.Context, payload *projsvc.Cr
 		return nil, domain.ErrInternal
 	}
 
-	projectFull := models.ConvertToProjectFull(projectDB, projectSettingsDB)
+	projectFull := ConvertToProjectFull(projectDB, projectSettingsDB)
 
 	slog.DebugContext(ctx, "returning created project", "project", projectFull)
 
@@ -204,7 +204,7 @@ func (s *ProjectsService) GetOneProjectBase(ctx context.Context, payload *projsv
 		return nil, domain.ErrInternal
 	}
 
-	project := models.ConvertToServiceProjectBase(projectDB)
+	project := ConvertToServiceProjectBase(projectDB)
 
 	// Store the revision in context for the custom encoder to use
 	revisionStr := strconv.FormatUint(revision, 10)
@@ -254,7 +254,7 @@ func (s *ProjectsService) GetOneProjectSettings(ctx context.Context, payload *pr
 		return nil, domain.ErrInternal
 	}
 
-	projectSettings := models.ConvertToServiceProjectSettings(projectSettingsDB)
+	projectSettings := ConvertToServiceProjectSettings(projectSettingsDB)
 
 	// Store the revision in context for the custom encoder to use
 	revisionStr := strconv.FormatUint(revision, 10)
@@ -378,7 +378,7 @@ func (s *ProjectsService) UpdateProjectBase(ctx context.Context, payload *projsv
 		project.CreatedAt = misc.StringPtr(existingProjectDB.CreatedAt.Format(time.RFC3339))
 	}
 
-	projectDB, err := models.ConvertToDBProjectBase(project)
+	projectDB, err := ConvertToDBProjectBase(project)
 	if err != nil {
 		slog.ErrorContext(ctx, "error converting project to DB project", constants.ErrKey, err)
 		return nil, domain.ErrInternal
@@ -427,7 +427,7 @@ func (s *ProjectsService) UpdateProjectBase(ctx context.Context, payload *projsv
 
 	slog.DebugContext(ctx, "returning updated project", "project", project)
 
-	projectResp := models.ConvertToServiceProjectBase(projectDB)
+	projectResp := ConvertToServiceProjectBase(projectDB)
 
 	return projectResp, nil
 }
@@ -505,7 +505,7 @@ func (s *ProjectsService) UpdateProjectSettings(ctx context.Context, payload *pr
 		projectSettings.CreatedAt = misc.StringPtr(existingProjectSettingsDB.CreatedAt.Format(time.RFC3339))
 	}
 
-	projectSettingsDB, err := models.ConvertToDBProjectSettings(projectSettings)
+	projectSettingsDB, err := ConvertToDBProjectSettings(projectSettings)
 	if err != nil {
 		slog.ErrorContext(ctx, "error converting project settings to DB project settings", constants.ErrKey, err)
 		return nil, domain.ErrInternal
